Move getOuterWidth to SimpleCarousel.utils
